extract dataclasses for business2018
